Obradjeni izuzeci

+++ backend/backend/Controllers/AddressController.cs
@@ -20,12 +20,20 @@ public class AddressController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpGet]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
     public async Task<ActionResult<IEnumerable<AddressDto>>> GetAllAddresses()
     {
         var addresses = await addressRepository.GetAllAddresss();
+        if (addresses == null || addresses.Count() == 0)
+        {
+            return NoContent();
+        }
         return Ok(addresses);
     }
 
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     [Authorize(Roles = "User,Admin")]
     [HttpGet("{id}")]
     public async Task<ActionResult<AddressDto>> GetAddressById(Guid id)
@@ -38,8 +46,12 @@ public class AddressController : ControllerBase
 
         return Ok(address);
     }
+
     [Authorize(Roles = "User")]
     [HttpPost]
+    [Consumes("application/json")]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<ActionResult<Guid>> CreateAddress(AddressCreateDto addressDto)
     {
         var addressId = await addressRepository.CreateAddress(addressDto);
@@ -48,8 +60,16 @@ public class AddressController : ControllerBase
 
     [Authorize(Roles = "User")]
     [HttpPut("{id}")]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> UpdateAddress(Guid id, AddressUpdateDto addressDto)
     {
+        var oldAddress = addressRepository.GetAddressById(id);
+        if(oldAddress  == null)
+        {
+            return NotFound();
+        }
         try
         {
             await addressRepository.UpdateAddress(id, addressDto);
@@ -63,8 +83,16 @@ public class AddressController : ControllerBase
 
     [Authorize(Roles = "User")]
     [HttpDelete("{id}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> DeleteAddress(Guid id)
     {
+        var address = addressRepository.GetAddressById(id);
+        if (address == null)
+        {
+            return NotFound();
+        }
         try
         {
             await addressRepository.DeleteAddress(id);
