+++ backend/backend/Controllers/AdminController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using backend.Data.Repository;
+using backend.Models;
 using backend.Models.Dtos;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -16,13 +17,23 @@ public class AdminController : ControllerBase
         this.adminRepository = adminRepository;
        this.mapper = mapper;
     }
+
     [Authorize(Roles = "Admin")]
     [HttpGet]
-    public async Task<ActionResult<IEnumerable<AdminDto>>> GetAllAdmines()
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    public async Task<ActionResult<IEnumerable<AdminDto>>> GetAllAdmins()
     {
-        var admines = await adminRepository.GetAllAdmins();
-        return Ok(admines);
+        var admins = await adminRepository.GetAllAdmins();
+        if (admins == null || admins.Count() == 0)
+        {
+            return NoContent();
+        }
+        return Ok(admins);
     }
+
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     [Authorize(Roles = "Admin")]
     [HttpGet("{id}")]
     public async Task<ActionResult<AdminDto>> GetAdminById(Guid id)
@@ -35,17 +46,33 @@ public class AdminController : ControllerBase
 
         return Ok(admin);
     }
+
+
+
     [Authorize(Roles = "Admin")]
     [HttpPost]
+    [Consumes("application/json")]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<ActionResult<Guid>> CreateAdmin(AdminCreateDto adminDto)
     {
         var adminId = await adminRepository.CreateAdmin(adminDto);
         return Ok(adminId);
     }
+
+
     [Authorize(Roles = "Admin")]
     [HttpPut("{id}")]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> UpdateAdmin(Guid id, AdminUpdateDto adminDto)
     {
+        var oldAdmin = adminRepository.GetAdminById(id);
+        if (oldAdmin == null)
+        {
+            return NotFound();
+        }
         try
         {
             await adminRepository.UpdateAdmin(id, adminDto);
@@ -56,10 +83,20 @@ public class AdminController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
+
     [Authorize(Roles = "Admin")]
     [HttpDelete("{id}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> DeleteAdmin(Guid id)
     {
+
+        var admin = adminRepository.GetAdminById(id);
+        if (admin == null)
+        {
+            return NotFound();
+        }
         try
         {
             await adminRepository.DeleteAdmin(id);
