+++ backend/backend/Controllers/DiscountController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using backend.Data.Repository;
+using backend.Models;
 using backend.Models.Dtos;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -18,12 +19,20 @@ public class DiscountController : ControllerBase
     }
     [Authorize(Roles = "Admin")]
     [HttpGet]
-    public async Task<ActionResult<IEnumerable<DiscountDto>>> GetAllDiscountes()
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    public async Task<ActionResult<IEnumerable<DiscountDto>>> GetAllDiscounts()
     {
-        var discountes = await discountRepository.GetAllDiscounts();
-        return Ok(discountes);
+        var discounts = await discountRepository.GetAllDiscounts();
+        if (discounts == null || discounts.Count() == 0)
+        {
+            return NoContent();
+        }
+        return Ok(discounts);
     }
 
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     [Authorize(Roles = "Admin,User")]
     [HttpGet("{id}")]
     public async Task<ActionResult<DiscountDto>> GetDiscountById(Guid id)
@@ -39,6 +48,9 @@ public class DiscountController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpPost]
+    [Consumes("application/json")]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<ActionResult<Guid>> CreateDiscount(DiscountCreateDto discountDto)
     {
         var discountId = await discountRepository.CreateDiscount(discountDto);
@@ -47,8 +59,16 @@ public class DiscountController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpPut("{id}")]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> UpdateDiscount(Guid id, DiscountUpdateDto discountDto)
     {
+        var oldDiscount = discountRepository.GetDiscountById(id);
+        if (oldDiscount == null)
+        {
+            return NotFound();
+        }
         try
         {
             await discountRepository.UpdateDiscount(id, discountDto);
@@ -62,8 +82,16 @@ public class DiscountController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpDelete("{id}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> DeleteDiscount(Guid id)
     {
+        var discount = discountRepository.GetDiscountById(id);
+        if (discount == null)
+        {
+            return NotFound();
+        }
         try
         {
             await discountRepository.DeleteDiscount(id);
