+++ backend/backend/Controllers/ProductConroller.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using backend.Data.Repository;
+using backend.Models;
 using backend.Models.Dtos;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -18,13 +19,21 @@ public class ProductController : ControllerBase
     }
 
     [HttpGet]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
     [AllowAnonymous]
     public async Task<ActionResult<IEnumerable<ProductDto>>> GetAllProducts()
     {
         var products = await productRepository.GetAllProducts();
+        if (products == null || products.Count() == 0)
+        {
+            return NoContent();
+        }
         return Ok(products);
     }
 
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     [HttpGet("{id}")]
     [AllowAnonymous]
     public async Task<ActionResult<ProductDto>> GetProductById(Guid id)
@@ -41,6 +50,9 @@ public class ProductController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpPost]
+    [Consumes("application/json")]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<ActionResult<Guid>> CreateProduct(ProductCreateDto productDto)
     {
         var productId = await productRepository.CreateProduct(productDto);
@@ -50,8 +62,16 @@ public class ProductController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpPut("{id}")]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> UpdateProduct(Guid id, ProductUpdateDto productDto)
     {
+        var oldProduct = productRepository.GetProductById(id);
+        if (oldProduct == null)
+        {
+            return NotFound();
+        }
         try
         {
             await productRepository.UpdateProduct(id, productDto);
@@ -66,8 +86,16 @@ public class ProductController : ControllerBase
 
     [Authorize(Roles = "Admin")]
     [HttpDelete("{id}")]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
     public async Task<IActionResult> DeleteProduct(Guid id)
     {
+        var product = productRepository.GetProductById(id);
+        if (product == null)
+        {
+            return NotFound();
+        }
         try
         {
             await productRepository.DeleteProduct(id);
